requirements.txt
Django>=5.1.2
psycopg2>=2.9.1
django-environ>=0.4.5
pillow>=10.4.0
django-cleanup>=9.0.0
djangorestframework>=3.15.2
djangorestframework-gis>=1.1
django-cors-headers>=4.4.0